+++ client/src/index.js
@@ -2,7 +2,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 
-import App from './components/App';
+import App from './App';
 import './index.css';
 
 ReactDOM.render(
